+++ src/main/java/insane96mcp/survivalreimagined/module/sleeprespawn/block/GraveBlockEntity.java
@@ -2,30 +2,33 @@ package insane96mcp.survivalreimagined.module.sleeprespawn.block;
 
 import insane96mcp.survivalreimagined.module.sleeprespawn.feature.Death;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.NonNullList;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.nbt.ListTag;
 import net.minecraft.network.protocol.Packet;
 import net.minecraft.network.protocol.game.ClientGamePacketListener;
 import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket;
-import net.minecraft.world.ContainerHelper;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.ArrayList;
+import java.util.List;
+
 public class GraveBlockEntity extends BlockEntity {
+    public static final String ITEMS_TAG = "items";
     public static final String XP_STORED_TAG = "xp_stored";
-    private NonNullList<ItemStack> items = NonNullList.create();
+    private List<ItemStack> items = new ArrayList<>();
     private int xpStored = 0;
     public GraveBlockEntity(BlockPos pos, BlockState state) {
         super(Death.GRAVE_BLOCK_ENTITY_TYPE.get(), pos, state);
     }
 
-    public NonNullList<ItemStack> getItems() {
+    public List<ItemStack> getItems() {
         return this.items;
     }
 
-    public void setItems(NonNullList<ItemStack> items) {
+    public void setItems(List<ItemStack> items) {
         this.items = items;
         this.setChanged();
     }
@@ -40,17 +43,27 @@ public class GraveBlockEntity extends BlockEntity {
     }
 
     @Override
-    public void load(CompoundTag p_155055_) {
-        super.load(p_155055_);
-        this.items = NonNullList.create();
-        ContainerHelper.loadAllItems(p_155055_, this.items);
-        this.xpStored = p_155055_.getInt(XP_STORED_TAG);
+    public void load(CompoundTag compoundTag) {
+        super.load(compoundTag);
+        ListTag itemsList = compoundTag.getList(ITEMS_TAG, 10);
+
+        for(int i = 0; i < itemsList.size(); ++i) {
+            CompoundTag itemStackTag = itemsList.getCompound(i);
+            this.items.add(ItemStack.of(itemStackTag));
+        }
+        this.xpStored = compoundTag.getInt(XP_STORED_TAG);
     }
 
     @Override
     protected void saveAdditional(CompoundTag compoundTag) {
         super.saveAdditional(compoundTag);
-        ContainerHelper.saveAllItems(compoundTag, this.items);
+        ListTag itemsList = new ListTag();
+        for (ItemStack itemStack : this.items) {
+            CompoundTag itemStackTag = new CompoundTag();
+            itemStack.save(itemStackTag);
+            itemsList.add(itemStackTag);
+        }
+        compoundTag.put(ITEMS_TAG, itemsList);
         compoundTag.putInt(XP_STORED_TAG, xpStored);
     }