+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Upcoming
+* Changed death (again)
+  * Dieing now summons a grave. Breaking the grave will give back your items and summon the players' ghost with the xp, like before.
 * Changed Beeg Veins feature to Ore Generation feature
   * Added Poor and Rich iron, gold and copper ores. The veins of these ores now generate randomly with poor, normal and rich ores
     * Poor ores drop 50% stuff while rich drops 200%